add ci workflow for testing/formatting
